renamed healthcheck test file and added swagger service tests
